edit to component
